refactor: getOrCreateBuildingId 로직을 어플리케이션 레이터로 이동
