+++ src/main/kotlin/com/wheretopop/application/building/BuildingFacade.kt
@@ -19,6 +19,13 @@ class BuildingFacade(
     }
 
     suspend fun getOrCreateBuildingId(command: BuildingCommand.CreateBuildingCommand): Long {
-        return buildingService.getOrCreateBuildingId(command)
+        val targetBuilding = buildingService.getBuilding(command.address)
+        if (targetBuilding == null) {
+            val building = this.createBuilding(command)
+            if (building == null) throw Exception("빌딩 생성에 실패했습니다.")
+            return building.id
+        } else {
+            return targetBuilding.id
+        }
     }
 }
