+++ src/main/kotlin/com/wheretopop/domain/building/BuildingService.kt
@@ -4,5 +4,4 @@ interface BuildingService {
     suspend fun getBuilding(address: String): BuildingInfo.Main?
     suspend fun searchBuildings(criteria: BuildingCriteria.SearchBuildingCriteria): List<BuildingInfo.Main>
     suspend fun createBuilding(command: BuildingCommand.CreateBuildingCommand): BuildingInfo.Main?
-    suspend fun getOrCreateBuildingId(command: BuildingCommand.CreateBuildingCommand): Long
 }
